Second assignment
Rubic Clock Task 10.

Create a game, which implements the Rubik clock. In this game there are 9 clocks. Each clock can
show a time between 1 and 12 (hour only). Clocks are placed in a 3x3 grid, and initially they set
randomly. Each four clocks on a corner has a button placed between them, so we have four buttons
in total. Pressing a button increase the hour on the four adjacent clocks by one. The player wins, if all
the clocks show 12.
Implement the game, and let the player restart it. The game should recognize if it is ended, and it has
to show in a message box how much steps did it take to solve the game. After this, a new game
should be started automatically.